+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/TransferManager.cs
@@ -71,11 +71,11 @@ namespace Azure.Storage.DataMovement
 
         /// <summary>
         /// Defines the error handling method to follow when an error is seen. Defaults to
-        /// <see cref="ErrorHandlingOptions.StopOnAllFailures"/>.
+        /// <see cref="ErrorHandlingBehavior.StopOnAllFailures"/>.
         ///
-        /// See <see cref="ErrorHandlingOptions"/>.
+        /// See <see cref="ErrorHandlingBehavior"/>.
         /// </summary>
-        internal ErrorHandlingOptions _errorHandling;
+        internal ErrorHandlingBehavior _errorHandling;
 
         /// <summary>
         /// Cancels the channels operations when disposing.
@@ -127,7 +127,7 @@ namespace Azure.Storage.DataMovement
             _checkpointer = _checkpointerOptions != default ? _checkpointerOptions.GetCheckpointer() : CreateDefaultCheckpointer();
             _dataTransfers = new Dictionary<string, DataTransfer>();
             _arrayPool = ArrayPool<byte>.Shared;
-            _errorHandling = options?.ErrorHandling != default ? options.ErrorHandling : ErrorHandlingOptions.StopOnAllFailures;
+            _errorHandling = options?.ErrorHandling != default ? options.ErrorHandling : ErrorHandlingBehavior.StopOnAllFailures;
         }
 
         #region Job Channel Management